Cache the results of readdirSync in loadModuleFromfile
When building feed-data, I found that this cut the readdir time in half,
saving ~5% of total build time.

It doesn't introduce a race - there already was one - but it makes it
more likely to be encountered, in the unlikely event that files are
being added and deleted *during* a build.